Made by @elonof
SimpleKatAim is a Roblox cheat script designed for the game "KAT" and created by user "elono". It features a silent aim that functions as an aimbot, allowing players to hit targets without manually aiming. Simply ensure that the player is within the circle on your screen and shoot, and the script will automatically hit them.